requirements.txt
Django==1.10.2
django-tables2==1.2.6
pep8==1.7.0
psycopg2==2.6.2
PyJWT==1.4.2
requests==2.12.5